// src/app/model/SetUserReq.model.ts
export class SetUserReqdModel{

	idUsuario?: number;
    nombre?: string;
    appPaterno?: string;
    appMaterno?: string;

    email?: string;
    celular?: string;
    idPais?: number;
    idComuna?: number;
    idRegion?: number;
    direccion?: string;
    idGenero?: number;
    
    constructor(){
    }
   
}